Remove album.title uniqueness constraint. Fixes #3
Requiring unique album titles doesn't seem to be a good idea as it's
possible multiple artists could have an album by the same title. This
commit requires that a unique title and artist be required in order to
create a new Album instance without representing that explicitly in the
database schema.